+++ packages/sdks/core-js-sdk/src/sdk/outbound/types.ts
@@ -1,4 +1,6 @@
 export type ConnectOptions = {
-  redirectURL?: string;
+  redirectUrl?: string;
   scopes?: string[];
+  tenantId?: string;
+  tenantLevel?: boolean;
 };