+++ packages/sdks/web-js-sdk/src/sdk/flow.ts
@@ -19,6 +19,7 @@ type Options = Pick<
   | 'locale'
   | 'oidcPrompt'
   | 'oidcErrorRedirectUri'
+  | 'oidcResource'
   | 'nativeOptions'
   | 'thirdPartyAppStateId'
   | 'applicationScopes'