+++ packages/widgets/outbound-applications-widget/src/lib/widget/mixins/initMixin/initMixin.ts
@@ -1,0 +1,19 @@
+import { compose, createSingletonMixin } from '@descope/sdk-helpers';
+import { debuggerMixin, themeMixin } from '@descope/sdk-mixins';
+import { flowRedirectUrlMixin } from '../flowRedirectUrlMixin';
+import { initOutboundAppsListMixin } from './initComponentsMixins/initOutboundAppsListMixin';
+
+export const initMixin = createSingletonMixin(
+  <T extends CustomElementConstructor>(superclass: T) =>
+    // @ts-ignore
+    class InitMixinClass extends compose(
+      debuggerMixin,
+      themeMixin,
+      flowRedirectUrlMixin, // This mixin must be before all other mixins that loads flows
+      initOutboundAppsListMixin,
+    )(superclass) {
+      async init() {
+        await super.init?.();
+      }
+    },
+);