+++ packages/widgets/outbound-applications-widget/src/lib/widget/state/selectors.ts
@@ -1,0 +1,32 @@
+import { createSelector } from 'reselect';
+import { State } from './types';
+
+export const getOutboundAppsList = (state: State) =>
+  state.outboundAppsList.data;
+
+export const getConnectedOutboundAppsList = (state: State) =>
+  state.connectedOutboundAppsIds.data;
+
+export const allowedAppsIds = (state: State) => state.allowedAppsIds.data;
+
+export const getAppsList = createSelector(
+  getOutboundAppsList,
+  getConnectedOutboundAppsList,
+  allowedAppsIds,
+  (obApps, connectedIds, allowedIds) =>
+    obApps
+      .filter((app) => allowedIds === undefined || allowedIds.includes(app.id))
+      .map((app) => {
+        const isConnected = connectedIds?.includes(app.id) || false;
+        return {
+          ...app,
+          appId: app.id,
+          isConnected,
+        };
+      })
+      .sort((a, b) => a.name.localeCompare(b.name)),
+);
+
+export const getMe = (state: State) => state.me.data;
+
+export const getUserId = createSelector(getMe, (me) => me.userId);