+++ packages/widgets/outbound-applications-widget/test/mocks/rootMock.ts
@@ -1,0 +1,5 @@
+export default `
+<descope-container data-editor-type="container" direction="column" id="ROOT" data-id="outbound-apps-list" space-between="md" st-horizontal-padding="1rem" st-vertical-padding="1rem" st-align-items="safe center" st-justify-content="safe center" st-host-width="100%" st-gap="1rem">
+  <descope-outbound-apps connect-flow-id="outbound-apps-connect" disconnect-flow-id="outbound-apps-disconnect" size="sm" />
+</descope-container>
+`;