+++ packages/widgets/role-management-widget/e2e/role-management-widget.spec.ts
@@ -1,4 +1,5 @@
 import { test, expect } from '@playwright/test';
+import { componentsPort, widgetPort } from '../playwright.config';
 import mockTheme from '../test/mocks/mockTheme';
 import { apiPaths } from '../src/lib/widget/api/apiPaths';
 import {
@@ -25,12 +26,12 @@ const MODAL_TIMEOUT = 500;
 
 test.describe('widget', () => {
   test.beforeEach(async ({ page }) => {
-    await page.addInitScript(() =>
+    await page.addInitScript((port) => {
       window.localStorage.setItem(
         'base.ui.components.url',
-        'http://localhost:8766/umd/index.js',
-      ),
-    );
+        `http://localhost:${port}/umd/index.js`,
+      );
+    }, componentsPort);
 
     await page.route('*/**/config.json', async (route) =>
       route.fulfill({ json: configContent }),
@@ -93,7 +94,9 @@ test.describe('widget', () => {
       }),
     );
 
-    await page.goto('http://localhost:5556', { waitUntil: 'networkidle' });
+    await page.goto(`http://localhost:${widgetPort}`, {
+      waitUntil: 'networkidle',
+    });
   });
 
   test('roles table', async ({ page }) => {